Pass pre-commit status to AjaxSource in options
This branch attempts to solve incorrect assignment of PUT method in
AjaxSource by saving the Model or Collection's .status property at
time of entry to Model.commit and Collection.commit to
options.preCommitStatus so that can be used in AjaxSource.commit
to determine proper method assignment. options.preCommitStatus is
set to undefined in AjaxSource.commit so as not to interfere with
subsequent commits.
See: http://forums.enyojs.com/discussion/2618/
- Initial commit with temp logging